+++ src/Symfony/Component/Console/Formatter/NullOutputFormatterStyle.php
@@ -23,11 +23,17 @@ final class NullOutputFormatterStyle implements OutputFormatterStyleInterface
 
     public function setBackground(string $color = null): void
     {
+        if (1 > \func_num_args()) {
+            trigger_deprecation('symfony/console', '6.2', 'Calling "%s()" without any arguments is deprecated, pass null explicitly instead.', __METHOD__);
+        }
         // do nothing
     }
 
     public function setForeground(string $color = null): void
     {
+        if (1 > \func_num_args()) {
+            trigger_deprecation('symfony/console', '6.2', 'Calling "%s()" without any arguments is deprecated, pass null explicitly instead.', __METHOD__);
+        }
         // do nothing
     }