+++ src/Symfony/Component/Console/Helper/Helper.php
@@ -25,6 +25,9 @@ abstract class Helper implements HelperInterface
 
     public function setHelperSet(HelperSet $helperSet = null)
     {
+        if (1 > \func_num_args()) {
+            trigger_deprecation('symfony/console', '6.2', 'Calling "%s()" without any arguments is deprecated, pass null explicitly instead.', __METHOD__);
+        }
         $this->helperSet = $helperSet;
     }