+++ src/Symfony/Component/Console/Input/InputArgument.php
@@ -95,6 +95,9 @@ class InputArgument
      */
     public function setDefault(string|bool|int|float|array $default = null)
     {
+        if (1 > \func_num_args()) {
+            trigger_deprecation('symfony/console', '6.2', 'Calling "%s()" without any arguments is deprecated, pass null explicitly instead.', __METHOD__);
+        }
         if ($this->isRequired() && null !== $default) {
             throw new LogicException('Cannot set a default value except for InputArgument::OPTIONAL mode.');
         }