+++ src/Symfony/Component/Form/FormConfigBuilder.php
@@ -340,6 +340,9 @@ class FormConfigBuilder implements FormConfigBuilderInterface
 
     public function setDataMapper(DataMapperInterface $dataMapper = null): static
     {
+        if (1 > \func_num_args()) {
+            trigger_deprecation('symfony/form', '6.2', 'Calling "%s()" without any arguments is deprecated, pass null explicitly instead.', __METHOD__);
+        }
         if ($this->locked) {
             throw new BadMethodCallException('FormConfigBuilder methods cannot be accessed anymore once the builder is turned into a FormConfigInterface instance.');
         }