+++ src/Symfony/Component/HttpFoundation/JsonResponse.php
@@ -79,6 +79,9 @@ class JsonResponse extends Response
      */
     public function setCallback(string $callback = null): static
     {
+        if (1 > \func_num_args()) {
+            trigger_deprecation('symfony/http-foundation', '6.2', 'Calling "%s()" without any arguments is deprecated, pass null explicitly instead.', __METHOD__);
+        }
         if (null !== $callback) {
             // partially taken from https://geekality.net/2011/08/03/valid-javascript-identifier/
             // partially taken from https://github.com/willdurand/JsonpCallbackValidator