+++ src/Symfony/Component/HttpKernel/HttpCache/ResponseCacheStrategy.php
@@ -110,7 +110,7 @@ class ResponseCacheStrategy implements ResponseCacheStrategyInterface
         $response->headers->set('Age', $this->age);
 
         if ($this->isNotCacheableResponseEmbedded) {
-            $response->setLastModified();
+            $response->setLastModified(null);
 
             if ($this->flagDirectives['no-store']) {
                 $response->headers->set('Cache-Control', 'no-cache, no-store, must-revalidate');