+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Mark `Symfony\Component\Workflow\Registry` as internal
+ * Deprecate calling `Definition::setInitialPlaces()` without arguments
 
 6.0
 ---